pyproject.toml
[project]
name = "components-agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "arxiv>=2.2.0",
    "crawl4ai>=0.6.3",
    "duckdb>=1.3.0",
    "langchain-community>=0.3.25",
    "langchain-google-genai>=2.1.5",
    "langchain-nvidia-ai-endpoints>=0.3.10",
    "langchain-openai>=0.3.22",
    "langchain-tavily>=0.2.4",
    "langchain-xai>=0.2.4",
    "langchain[google-genai,groq,openai]>=0.3.25",
    "langgraph>=0.4.8",
    "notebook>=7.4.3",
    "pymongo>=4.13.1",
    "pymupdf>=1.26.1",
    "python-dotenv>=1.1.0",
    "rich>=14.0.0",
    "streamlit>=1.45.1",
    "streamlit-extras>=0.7.1",
    "textual>=3.4.0",
    "textual-dev>=1.7.0",
]
